Update DatabaseGUI.java with better column spacing
Added a new function called getLongestColumnNameLength(). It finds the length of the longest column of the output. Added new code segment in executeQuery to implement the new function. The new code spaces the columns based on the largest one. This code needs to be updated to add less space between thin columns.